+++ src/make.jl
@@ -11,7 +11,8 @@ The model/system is _not_ structurally simplified.
    or not having limitations on the left-hand-side (LHS) form.
 1. An `Equation`. The LHS format of the equation is limited.
    Let `x` be a `@variable` and `p` be a `@parameter`. Then, the LHS can only be one of:
-   `x`, `Differential(t)(x)`, `Differential(t)(x)*p`, `p*Differential(t)(x)`.
+   `x`, `Differential(t)(x)`, `Differential(t)(x)*p`, `p*Differential(t)(x)`,
+   however, the versions with `p` may fail unexpectedly.
    Anything else will either error or fail unexpectedly.
 2. A vector of the above two, which is then expanded. This allows the convenience of
    functions representing a physical process that may require many equations to be defined.