+++ CHANGELOG.md
@@ -16,4 +16,4 @@ parameters, and default processes.
 - New keyword `warn_default` in `processes_to_mtkmodel`.
 - Now `processes_to_mtkmodel` allows for `XDESystems` as elements of the input vector.
 - Detection of LHS-variable in `Equation` has been improved significantly.
-  Now, LHS can be any of: `x`, `Differential(t)(x)`, `Differential(t)(x)*p`, `p*Differential(t)(x)`.
+  Now, LHS can be any of: `x`, `Differential(t)(x)`, `Differential(t)(x)*p`, `p*Differential(t)(x)`. However, the multiplication versions often fail for unknown reasons.